+++ CHANGES.md
@@ -1,5 +1,7 @@
 ## Changes in version 0.4.5 (in development)
 
+* Default colour for dataset-related place groups now red while 
+  initial user places colour is colour. (#153)
 
 ## Changes in version 0.4.4